Eliminate data exports
Update Upa IDNA to version 2.3.0, which also eliminated data exports.

This makes the shared library easier to use, especially on Windows.
