+++ CMakeLists.txt
@@ -133,7 +133,6 @@ if (UPA_BUILD_TESTS OR UPA_BUILD_BENCH OR UPA_BUILD_FUZZER OR UPA_BUILD_EXAMPLES
       src/public_suffix_list.cpp
       src/url.cpp
       src/url_ip.cpp
-      src/url_percent_encode.cpp
       src/url_search_params.cpp
       src/url_utf.cpp)
     target_include_directories(${upa_lib_target} PUBLIC
