+++ LICENSE
@@ -1,6 +1,6 @@
 BSD 2-Clause License
 
-Copyright (c) 2016-2024 Rimas Misevičius
+Copyright (c) 2016-2025 Rimas Misevičius
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
@@ -25,8 +25,8 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 -------------------------------------------------------------------------------
 
-Files buffer.h, url.h, url_percent_encode.cpp, url_percent_encode.h contains
-portions of modified code from the Chromium project licensed as follows:
+Files buffer.h, url.h, url_percent_encode.h contains portions of modified code
+from the Chromium project licensed as follows:
 
 // Copyright 2015 The Chromium Authors. All rights reserved.
 //
