+++ examples/url-demo/build-url-api.sh
@@ -12,7 +12,6 @@ emcc examples/url-demo/url-api.cpp \
   src/public_suffix_list.cpp \
   src/url.cpp \
   src/url_ip.cpp \
-  src/url_percent_encode.cpp \
   src/url_search_params.cpp \
   src/url_utf.cpp \
   src/idna.cpp \
