+++ include/upa/url_percent_encode.h
@@ -342,7 +342,10 @@ constexpr bool is_ascii_alpha(CharT ch) noexcept {
 
 // Maps the hex numerical values 0x0 to 0xf to the corresponding ASCII digit
 // that will be used to represent it.
-extern UPA_API const char kHexCharLookup[0x10];
+inline constexpr char kHexCharLookup[0x10] = {
+    '0', '1', '2', '3', '4', '5', '6', '7',
+    '8', '9', 'A', 'B', 'C', 'D', 'E', 'F',
+};
 
 // This lookup table allows fast conversion between ASCII hex letters and their
 // corresponding numerical value. The 8-bit range is divided up into 8
@@ -352,7 +355,16 @@ extern UPA_API const char kHexCharLookup[0x10];
 // the corresponding numerical value.
 //
 // See hex_char_to_num for the lookup.
-extern UPA_API const char kCharToHexLookup[8];
+inline constexpr char kCharToHexLookup[8] = {
+    0,         // 0x00 - 0x1f
+    '0',       // 0x20 - 0x3f: digits 0 - 9 are 0x30 - 0x39
+    'A' - 10,  // 0x40 - 0x5f: letters A - F are 0x41 - 0x46
+    'a' - 10,  // 0x60 - 0x7f: letters a - f are 0x61 - 0x66
+    0,         // 0x80 - 0x9F
+    0,         // 0xA0 - 0xBF
+    0,         // 0xC0 - 0xDF
+    0,         // 0xE0 - 0xFF
+};
 
 // Assumes the input is a valid hex digit! Call is_hex_char before using this.
 inline unsigned char hex_char_to_num(unsigned char c) noexcept {
