--- src/url_percent_encode.cpp
@@ -1,33 +1,0 @@
-// Copyright 2016-2024 Rimas Misevičius
-// Distributed under the BSD-style license that can be
-// found in the LICENSE file.
-//
-// This file contains portions of modified code from:
-// https://cs.chromium.org/chromium/src/url/url_canon_internal.cc
-// Copyright 2013 The Chromium Authors. All rights reserved.
-//
-
-#include "upa/url_percent_encode.h"
-
-namespace upa::detail {
-
-// Maps hex numerical values to ASCII digits
-const char kHexCharLookup[0x10] = {
-    '0', '1', '2', '3', '4', '5', '6', '7',
-    '8', '9', 'A', 'B', 'C', 'D', 'E', 'F',
-};
-
-// Used to convert ASCII hex digit to numerical value
-const char kCharToHexLookup[8] = {
-    0,         // 0x00 - 0x1f
-    '0',       // 0x20 - 0x3f: digits 0 - 9 are 0x30 - 0x39
-    'A' - 10,  // 0x40 - 0x5f: letters A - F are 0x41 - 0x46
-    'a' - 10,  // 0x60 - 0x7f: letters a - f are 0x61 - 0x66
-    0,         // 0x80 - 0x9F
-    0,         // 0xA0 - 0xBF
-    0,         // 0xC0 - 0xDF
-    0,         // 0xE0 - 0xFF
-};
-
-
-} // namespace upa::detail
