+++ src/url_search_params.cpp
@@ -1,12 +1,14 @@
-// Copyright 2016-2023 Rimas Misevičius
+// Copyright 2016-2025 Rimas Misevičius
 // Distributed under the BSD-style license that can be
 // found in the LICENSE file.
 //
 
+#include "upa/url_percent_encode.h"
 #include "upa/url_search_params.h"
 
 
 namespace upa {
+namespace {
 
 // The percent encoding/mapping table of URL search parameter bytes. If
 // corresponding entry is '%', then byte must be percent encoded, otherwise -
@@ -14,7 +16,7 @@ namespace upa {
 // The table is based on:
 // https://url.spec.whatwg.org/#concept-urlencoded-serializer
 
-const char url_search_params::kEncByte[0x100] = {
+const char kEncByte[0x100] = {
 //   0    1    2    3    4    5    6    7    8    9    A    B    C    D    E    F
     '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', // 0
     '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', // 1
@@ -33,5 +35,20 @@ const char url_search_params::kEncByte[0x100] = {
     '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', // E
     '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%', '%'  // F
 };
+
+} // namespace
+
+void url_search_params::urlencode_sv(std::string& encoded, string_view value)
+{
+    for (const char c : value) {
+        const auto uc = static_cast<unsigned char>(c);
+        const char cenc = kEncByte[uc];
+        encoded.push_back(cenc);
+        if (cenc == '%') {
+            encoded.push_back(detail::kHexCharLookup[uc >> 4]);
+            encoded.push_back(detail::kHexCharLookup[uc & 0xF]);
+        }
+    }
+}
 
 } // namespace upa
