+++ tools/amalgamate/config-cpp.json
@@ -5,7 +5,6 @@
     "src/idna.cpp",
     "src/url.cpp",
     "src/url_ip.cpp",
-    "src/url_percent_encode.cpp",
     "src/url_search_params.cpp",
     "src/url_utf.cpp"
   ],
